Allow dynamic updates in the GUI
This means that updates from the robot_description topic are
properly shown in the UI now.  While I was in here, I also
changed it so that the UI "flows" as you resize the window.
This should be much more usable on smaller screens.  It
also simplifies the code quite a bit.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>