Port to ROS 2 (#30)
* Port to ROS 2.

* Add in copyright and XML lint tests.

* Allow dynamic updates in the GUI

This means that updates from the robot_description topic are
properly shown in the UI now.  While I was in here, I also
changed it so that the UI "flows" as you resize the window.
This should be much more usable on smaller screens.  It
also simplifies the code quite a bit.

* Initialize gui sliders at startup

* Add 64_joint_robot launch test
* Add mimic cycle launch test
* Add launch test for mimic chain
* Add test comparing urdf and collada joints
* Add zero joint launch test

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>