Fix the mimic_cycle test to always succeed. (#74)
* Fix the mimic_cycle test to always succeed.

In particular, we weren't keeping the process alive, and we
weren't waiting for the process to finish before checking the
output.  Because of that, the test only worked if the launched
process happened to complete before the checks fired.

Instead, mark things as keep_alive, and make sure to have a
concurrent test that waits for the process to finish before
doing checks.  This makes the test always finish.

Also add a missing test_depend on ros2topic.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>