Clean up parsing to handle errors much better.
It is very possible to start parsing a URDF or
COLLADA file, and for an error to occur during that
parsing.  Previously, if this happened we would leave
the object in an undefined state with respect to the
free_joints, joint_list, and dependent_joints.

With this commit, we do all of the parsing and collecting
of data at once.  If that is completely successful, then
we update the object state to include that, so we can
no longer get partial parses.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>