Cleanup COLLADA support (#96)
* Clean up parsing to handle errors much better.

It is very possible to start parsing a URDF or
COLLADA file, and for an error to occur during that
parsing.  Previously, if this happened we would leave
the object in an undefined state with respect to the
free_joints, joint_list, and dependent_joints.

With this commit, we do all of the parsing and collecting
of data at once.  If that is completely successful, then
we update the object state to include that, so we can
no longer get partial parses.

* Refactor names to be more descriptive.

The code should be a little easier to follow this way.

* Additional error checking for COLLADA files.

* Unindent init_collada loop.

This just makes it easier to read.

* Always create the subscription to 'robot_description'.

That way it can be dynamically updated, even when the
initial configuration was given on the command-line.

* Make error messages for joint limits much more specific.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>